Move NFT Txs to My NFT page
